Added case_mixin and variant_mixin
