+++ puzzle/plugins/gemini/__init__.py
@@ -1,1 +1,3 @@
+from .case_mixin import CaseMixin
+from .variant_mixin import VariantMixin
 from .gemini_plugin import GeminiPlugin
